do not fail terribly on teardown when ldap could not be set up

+++ tests/Zend/Ldap/Node/UpdateTest.php
@@ -48,6 +48,8 @@ class Zend_Ldap_Node_UpdateTest extends Zend_Ldap_OnlineTestCase
 
     protected function tearDown()
     {
+        if(!$this->_getLdap()) return;
+
         foreach ($this->_getLdap()->getBaseNode()->searchChildren('objectClass=*') as $child) {
             $this->_getLdap()->delete($child->getDn(), true);
         }
